+++ .gitignore
@@ -25,7 +25,6 @@ build/Release
 dist/
 
 node_modules/*
-!node_modules/_app
 
 Factorio*
 /factorio
