+++ lib/clusterTools/checkHotpatchInstallation.js
