+++ lib/clusterTools/getCommand.js
