+++ lib/clusterTools/getInstanceName.js
@@ -32,7 +32,7 @@ module.exports = function(config){
 		}
 	} else {
 		return async function thisAintGonnaWork(){
-			let err = new Error("getInstanceName from _app/clusterTools needs to be called with {config} to work");
+			let err = new Error("getInstanceName from lib/clusterTools needs to be called with {config} to work");
 			console.log(err);
 			throw err;
 		}
