+++ lib/factorioServer/index.js
@@ -17,7 +17,7 @@ const path = require("path");
 const mkdirp = require("mkdirp");
 const appRootPath = require("app-root-path");
 
-const fileOps = require("_app/fileOps");
+const fileOps = require("lib/fileOps");
 
 module.exports = class factorioServer {
 	constructor({
