+++ lib/fileOps/fileOps.js
