+++ lib/fileOps/fileOps.spec.js
