+++ lib/fileOps/index.js
