+++ lib/generateSSLcert.spec.js
@@ -1,7 +1,7 @@
 var assert = require("assert");
 var fs = require("fs");
 
-var fileOps = require("_app/fileOps");
+var fileOps = require("lib/fileOps");
 var generateSSLcert = require("./generateSSLcert.js")
 
 describe("generateSSLcert.js(options)", ()=>{
