+++ lib/getInstances.js
@@ -1,4 +1,4 @@
-const fileOps = require("_app/fileOps");
+const fileOps = require("fileOps");
 // const config = require("./../config");
 
 module.exports = function(config) {
