+++ lib/isFactorioCommand/index.js
