+++ lib/isFactorioCommand/isFactorioCommand.js
