+++ lib/isFactorioCommand/isFactorioCommand.spec.js
