+++ lib/manager/modManager.js
@@ -13,7 +13,7 @@ const factorioAPI = require("node-factorio-api");
 const sLog = require("single-line-log").stdout;
 const query = require('cli-interact').getYesNo;
 
-const fileOps = require("_app/fileOps");
+const fileOps = require("lib/fileOps");
 // const config = require("./../../config");
