+++ lib/package.json
@@ -1,0 +1,5 @@
+{
+    "name": "lib",
+    "description": "Internal library for factorioClusterio"
+}
+
