+++ sharedPlugins/statisticsExporter/index.js
@@ -2,7 +2,7 @@ const needle = require("needle");
 const fs = require("fs-extra");
 const path = require("path");
 
-const clusterTools = require("_app/clusterTools")();
+const clusterTools = require("lib/clusterTools")();
 const pluginConfig = require("./config");
 
 module.exports = class remoteCommands {
